requirements.txt
Flask==3.0.2
opencv-python==4.9.0.80
numpy==1.26.4
mysql-connector-python==8.3.0
face-recognition==1.3.0
Pillow==10.2.0 